requirements.txt
customtkinter==5.2.2
pillow==10.2.0
python-nmap==0.7.1
